Add permission check to controllers

See merge request clear-code/redmine-plugin-testcase-management!12

+++ app/controllers/test_plans_controller.rb
@@ -2,6 +2,12 @@ class TestPlansController < ApplicationController
 
   include ApplicationsHelper
 
+  before_action :find_project_id
+  before_action :find_test_plan, :only => [:show, :edit, :update, :destroy]
+  before_action :find_test_plan_id, :only => [:assign_test_case, :unassign_test_case]
+  before_action :find_test_case_id, :only => [:unassign_test_case]
+  before_action :authorize_with_issues_permission, :except => [:index, :new, :create, :assign_test_case, :unassign_test_case]
+
   before_action do
     prepare_issue_status_candidates
     prepare_user_candidates
@@ -9,51 +15,21 @@ class TestPlansController < ApplicationController
 
   # GET /projects/:project_id/test_plans
   def index
-    begin
-      find_project(params.permit(:project_id)[:project_id])
-      @test_plans = TestPlan.all
-    rescue
-      flash.now[:error] = l(:error_project_not_found)
-      render 'forbidden', status: 404
-    end
+    @test_plans = TestPlan.where(project_id: @project.id).visible
   end
 
   # GET /projects/:project_id/test_plans/:id
   def show
-    begin
-      find_project(params.permit(:project_id)[:project_id])
-      begin
-        @test_plan = TestPlan.find(params.permit(:project_id, :id)[:id])
-        @test_case_test_plan = TestCaseTestPlan.new
-      rescue
-        flash.now[:error] = l(:error_test_plan_not_found)
-        render 'forbidden', status: 404
-      end
-    rescue
-      flash.now[:error] = l(:error_project_not_found)
-      render 'forbidden', status: 404
-    end
+    @test_case_test_plan = TestCaseTestPlan.new
   end
 
   # GET /projects/:project_id/test_plans/:id/edit
   def edit
-    begin
-      find_project(params.permit(:project_id)[:project_id])
-      begin
-        @test_plan = TestPlan.find(params.permit(:id)[:id])
-      rescue
-        flash.now[:error] = l(:error_test_plan_not_found)
-        render 'forbidden', status: 404
-      end
-    rescue
-      flash.now[:error] = l(:error_project_not_found)
-      render 'forbidden', status: 404
-    end
   end
 
   # PUT /projects/:project_id/test_plans/:id
   def update
-    @test_plan = TestPlan.find(params.permit(:id)[:id])
+    raise ::Unauthorized unless @test_plan.editable?
     update_params = {}
     update_params[:name] = test_plan_params[:name]
     update_params[:begin_date] = test_plan_params[:begin_date]
@@ -83,7 +59,9 @@ class TestPlansController < ApplicationController
 
   # POST /projects/:project_id/test_plans
   def create
-    find_project(params.permit(:project_id)[:project_id])
+    unless User.current.allowed_to?(:add_issues, @project, :global => true)
+      raise ::Unauthorized
+    end
     @test_plan = TestPlan.new(:name => test_plan_params[:name],
                               :begin_date => test_plan_params[:begin_date],
                               :end_date => test_plan_params[:end_date],
@@ -102,33 +80,29 @@ class TestPlansController < ApplicationController
 
   # DELETE /projects/:project_id/test_plans/:id
   def destroy
+    raise ActiveRecord::RecordNotFound unless @test_plan.visible?
+    raise ::Unauthorized unless @test_plan.deletable?
     begin
-      find_project(params.permit(:project_id)[:project_id])
-      begin
-        @test_plan = TestPlan.find(params.permit(:id)[:id])
-        if @test_plan.destroy
-          flash[:notice] = l(:notice_successful_delete)
-          redirect_to project_test_plans_path
-        else
-          flash.now[:error] = l(:error_delete_failure)
-          render :show
-        end
-      rescue
-        flash.now[:error] = l(:error_test_plan_not_found)
-        render 'forbidden', status: 404
+      if @test_plan.destroy
+        flash[:notice] = l(:notice_successful_delete)
+        redirect_to project_test_plans_path
+      else
+        flash.now[:error] = l(:error_delete_failure)
+        render :show
       end
     rescue
-      flash.now[:error] = l(:error_project_not_found)
+      flash.now[:error] = l(:error_test_plan_not_found)
       render 'forbidden', status: 404
     end
   end
 
   # POST /projects/:project_id/test_plans/:test_plan_id/assign_test_case
   def assign_test_case
+    return unless authorize_with_issues_permission(params[:controller], :create)
     begin
-      find_project(params.permit(:project_id)[:project_id])
-      @test_plan = TestPlan.find(params.permit(:test_plan_id)[:test_plan_id])
       @test_case = TestCase.find(params.require(:test_case_test_plan).permit(:test_case_id)[:test_case_id])
+      raise ActiveRecord::RecordNotFound unless @test_case.visible?
+      raise ActiveRecord::RecordNotFound unless @test_plan.visible?
       @test_case_test_plan = TestCaseTestPlan.where(test_plan: @test_plan,
                                                     test_case: @test_case).first
       unless @test_case_test_plan
@@ -146,10 +120,10 @@ class TestPlansController < ApplicationController
 
   # DELETE /projects/:project_id/test_plans/:test_plan_id/assign_test_case/:test_case_id
   def unassign_test_case
+    return unless authorize_with_issues_permission(params[:controller], :destroy)
     begin
-      find_project(params.permit(:project_id)[:project_id])
-      @test_plan = TestPlan.find(params.permit(:test_plan_id)[:test_plan_id])
-      @test_case = TestCase.find(params.permit(:test_case_id)[:test_case_id])
+      raise ActiveRecord::RecordNotFound unless @test_case.visible?
+      raise ActiveRecord::RecordNotFound unless @test_plan.visible?
       @test_case_test_plan = TestCaseTestPlan.where(test_plan: @test_plan,
                                                     test_case: @test_case).first
       if @test_case_test_plan
